Test `executeMUDRNRAPU001129` moved into `MULANGT18ProcedureTest`
